sync with no sync token and limit=0 when logging in on a matrix server
Perform an initial sync when logging in on a matrix server: Use no sync
token and limit events per room to 0, so all rooms are retrieved from
the matrix server. New events since the last run are retrieved later by
the listener thread with the sync token loaded from the accounts file.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>